This is a tool for rewriting import URIs.
Unfortunately, when using ozl to link functors, all import
URIs in the resulting functor have been resolved, i.e., in
the case of file names, are absolute.  This tool will help
convert then into relative URIs again, which we need for
installing.

We might want to have such functionality (in this or a
different form integrated into ozl.  We will propose this
shortly after the upcoming Mozart release, should our
experience with this tool for installing Stockhausen
prove it valuable.